Test: Turn test traceback-related variables to properties
To avoid accidental write into test-traceback-related variables, let's
turn them into properties. This makes it harder to manually fail the
test with custom exceptions, but I consider that ugly and one should
simply raise the exception and the runner should take care of setting
the class, traceback and details.

To prove the direct access is not necessary I tried this with
Avocado-vt which works without any modifications even though it uses
custom exception handling.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>