Test: Turn `params` into a property
This change should decrease the probability of users accidentally
overriding `Test.params` while keeping the usage the same. There is one
caveat and that is Avocado-vt, which modifies `self.params`, which is
not that simple anymore. I have a fix for it, but the problem is it is
not backward compatible and it can only be changed to
forward-compatible. We should consider whether the benefit of having
this "a bit" protected is greater than backward incompatible change with
Avocado-vt (and possibly other "nasty" user's tests)

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>