Test: Avoid sharing default_params between instances
The `default_params` is a dictionary, which is mutable object, therefor
it will be shared across all instances. It should not affect our code
much as we import the code in separate processes, but still it's not
nice keeping this in our code (as it might change in the future).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>