Test: Stop publishing non-existing values in get_state
We removed several classes in the history, let's update the get_state to
remove methods, that are not used anymore.

Note that some of the variables were related to Avocado-vt, anyway the
usage of `get_state` from Avocado depends only on it's values. If
Avocado-vt needs to publish it's internal values, it can override this
method and do so, anyway at this point this is not necessary.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>