Test: Turn runner_queue into a property
The runner queue is essential part of runner->test communication. Let's
decrease the probability of user accidentally replacing the runner queue
by turning it into a property.

For safety reasons our runner also needs to set this property from
outside (if it's not set by the test_loader) so this patch also includes
a method `set_runner_queue` which allows overriding the runner queue
from outside, but we only allow this once to avoid unwanted changes.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>