varianter: Remove the unnecessary "ignore_new_data" interface
The "ignore_new_data" was used by "replay" to override Varianter and
tell it to stop accepting new data by other plugins (like yaml_to_mux).
With the new changes the plugins are part of the Varianter, so when
"replay" replaces it with the old (parsed) one, it will not parse itself
again, therefor the default_params and variants stays untouched.

To make sure this works well a unittest is included.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>