varianter: Create the concept of default_params
The idea behind default_params is, that some of the params are
considered the default_params, they can change, but are not part of
variant, nor the variant-id.

This implementation replaces the `Mux.data_inject` with
`Mux.add_default_param` which also accepts `name` to map data to
different plugins/groups in order to allow ignoring only some of them.

During parsing it extends multiplex variants of the defaults (merges
multiplex variants into defaults, therefor the mux overrides the values)
in order to provide the combination while yielding variants. Then when
there are variants, it yields the variant (compound of the variant and
the default_params) and variant-id. If there are no variants, it yields
only the default_params with no variant-id (None).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>