varianter: Make it properly pluggable
This patch replaces the hardcoded mux in Varianter for list of Varianter
plugins, iterating through all of them to produce all specified
variants. The interface is described in `avocado.core.plugin_interfaces`
and we also provide base-implementation for Mux-like plugins in
`avocado.core.mux`.

Mainly this PR is about defining the stevedore entry points, moving the
code around and wrapping the code into a class according the defined
stevedore interface. Doing so allowed us to simplify the `MuxTree`
object which does not need to keep track of the `root` of the tree,
because the root is stored in the pluggin's instance.

There is still one potential issue in Varianter which is probably even
more visible now, which is that we are pickling it in jobdata in order
to be able to reproduce the variants when re-running/diffing the job.
Before this was prone to code-changes but now it can be even worse as
multiple plugins can be enabled and can break loading of the previous
state. To avoid the potential breakage this version simply initializes
only the available extensions which should produce only the supported
variants.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>