runner: fix _iter_variants()
variants.itertests() changed the format of each variant from a tuple to
a dictionary. After that change, the runner was not updated to provide
the mux_path considering the new format.

This patch adds the mux_path to the factory params in _iter_variants(),
fixing the Test.self.__params resolution.

Signed-off-by: Amador Pahim <apahim@redhat.com>