RPM package rename: from avocado to python-avocado
This is a step towards reducing the delta between Avocado "upstream"
packages, that is, the packages offered in our own custom repo, and
packages provided by "dowsntream" distros such as Fedora and (EP)EL.

The package rename is probably the change that could affect users
the most, so let's try to do this early (sprint-wise).  The proper
"Provides" and "Obsoletes" tags should, though, make the update
seamless.

Signed-off-by: Cleber Rosa <crosa@redhat.com>