Don't accept skip decorators in setUp() and tearDown()
Skip decorators are meant to skip a test. Using them in
setUp()/tearDown() can potentially mislead users, making them to believe
that only the decorated method will be skipped, but not the test method,
while in fact the test method will also be skipped.

Let's drop this magic behaviour and accept the skip decorators to be
used only in the test methods, making the test to ERROR if used
somewhere else.

Signed-off-by: Amador Pahim <apahim@redhat.com>