mux: Add support for internal filters
This adds the support to define per-node filters, which take part of
environment and are evaluated during multiplexation.

The basic idea is to define relations ({could be only used | could
not be used} with other branch), which makes it a bit different from
what we know from `--mux-filter-{only|out}`:

1. They are applied per complete variant (because the filters are
   compound of per-leaf-filters)
2. They don't remove/keep the nodes, but they only evaluate whether
   this resulting variant makes sense (whether it passes the filters)

See the updated documentation for details.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>