Optional plugins: do not depend on the VERSION file
Simply because this file won't be available when being installed
from any method outside the source tree.

The downside is obvious: it requires extra steps when doing a new
release.  But that is still better than not being able to ship
these plugins to users using the standard Python mechanisms.

Signed-off-by: Cleber Rosa <crosa@redhat.com>