selftests: Run check_tmp_dirs between selftests
This commit enables the left-behind-temporary-dirs check between each
test. As all selftests are executed as a single process the BORG class
keeps the TMP dir of the unit-tests. To avoid results spoiled by this
fact this commit adds a force-cleanup of the tmp dir tracker overriding
the BORG internal state, which is not really nice, but for selftests
acceptable.

The method to clean it was especially named
"unittest_refresh_dir_tracker" to emphasize it should not be used
outside of unittest.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>