utils.archive: Skip extra treatment of existing symlinks
When one extracts a zip file with symlinks multiple times, the symlinks
are not replaced by the `extractall`, which results in our
`_update_zip_extra_attrs` workaround to try to create a symlink to the
content of the file, rather than to the destination. This patch checks
whether the link already exists, removes it and refresh it from the
zipfile.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>