selftest: Avoid marking non-sensitive test as sensitive
This test was marked as timing-sensitive test, but in reality it should
finish immediately (0.5s). In case of regression this unveils a
fork-bomb which could break the testing. That is the reason for the
timeout which was too-tight for some environments so let's increase the
timeout which will be only used on Avocado regression.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>