html: Avoid removing in-job-dir results on --html
When --html is used the original in-job-dir results are removed just
before the second render. This was probably due to typo as the original
location was (probably) intended for removal. This patch doesn't do that
because the user-defined dir can be shared dir with other user files so
let's just produce the results and keep the dirs intact (while keeping
the refresh in job-dir as that is guaranteed to be our directory and we
do know it can be removed).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>