replay: Include non-executed tests in replay map
The "results.json" includes only executed tests in "tests" while the
non-executed tests are only included in the overall info like "total".
Let's append "UNKNOWN" tests at the end of the "tests" to represent
those non-executed tests and mark them as "INTERRUPTED".

According to our meanings they'd be "SKIP" tests, but this is local
representation for Replay only and we would not be able to distinguish
between SKIP and SKIP-AFTER-INTERRUPTION, which is super handy. To avoid
introducing new/artifical status, let's use INTERRUPTED which works for
gur scenario and in the future we might expand it if needed.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>