selftests: introduce multiple levels of testing coverage
A finer level of granularity may help to extend our testing coverage
while reducing the amount of false positives.  This introduces the
three different levels, mapped like this:

   * level 0, AKA "make check"
   * level 1, AKA "make check-extended"
   * level 2, AKA "make check-full"

There are no changes of test assignment, that is, tests previously
being run under "check-full" will continue to be run only at that
target.  The same is true for tests that would run under "make check"
before this.

The big change is that there's now a middle ground, "make
check-extended", that can be used for environments that are halfway
between a dedicated machine and a really low powered environment.

Signed-off-by: Cleber Rosa <crosa@redhat.com>