Job: add time accounting
It may be useful to users of the Job class to have information on how
much time it took for it to run.  Since the goal for the Job class is
to be flexible when it comes to its usage, it may not be possible to
track the execution time if users follow a different path.

For now, the time accouting is done automatically if users use the
`run()` method.

If advanced users of the Job class choose to set the start time, end
time and elapsed time themselves, they are free to do it and `run()`
will never overwrite it.

Signed-off-by: Cleber Rosa <crosa@redhat.com>