replay: Add support to resume interrupted/crashed job
This commit adds "--replay-resume" argument, which is basically a
symlink to "--replay-test-status INTERRUPTED", therefor it executes only
the tests which were interrupted, or not executed (as we use INTERRUPTED
in Replay internally to represent not-executed-tests).

As the "results.json" used to get the list of tests might not be
available on a system crash, this also adds a fallback to use
"results.tap", which is a machine readable, streamline format. It does
not represent all available test statuses, but as a fallback it works
well.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>