avocado: Make LOG_UI and LOG_JOB available in public API
Currently on many places (mainly in plugins) we depend on hardcoded log
location to make UI and JOB/TEST loggers available to developers. This
is not really correct and it makes sense to make this available in the
standard API.

The "LOG_UI" is actually "avocado.app" logger and the "LOG_JOB" is
"avocado.test". I think those names fit better the usage as the APP
logger is used for UI and TEST logger exists and is used before and
after the test execution to also inform about job-related stuff.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>