output: Define LOG_UI and LOG_JOB
On many places we are directly getting "avocado.app" and "avocado.test"
loggers and interacting with them. This can become an issue when we try
to replace or modify our logging streams in the future as well as it
requires some hidden knowledge. Let's define LOG_UI and LOG_JOB in
`avocado.core.output` to be available for Avocado core (and silently
also for plugin) developers.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>