Human plugin: only print job results when job is successful
The UI was showing information about the job, no matter if the job got
to run tests.

Even though the formal available statuses of a job are not too evolved
and documented, Avocado already sets a job status as "PASS" when it
finishes running the tests.  This involves most of the job phases.

Let's use that to check if the UI should print information about the
test results and job execution time.

Signed-off-by: Cleber Rosa <crosa@redhat.com>