output: Support Paginator.flush
While debugging Avocado I noticed some crashes when attaching debugger.
It was caused by the debugger which force-flushes the STDOUT on some
occasions, which is usually supported, but is not supported by our
Paginator class. This commit adds the flush method which flushes the
output pipe (unless it's already closed)

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>