Revert "optional_plugins.robot: Report failures on verbose list"
I honestly missed the fact that no other loader reports failures
when it doesn't find resolve a test for a given test reference.
This brings experience that is really inconsistent.

Since the real fix may not be so simple to come up with, I propose we
revert this commit at this point.

This reverts commit 2b52631a4d51a7e7a84f66cb7c05d2b186c4edd9.

Signed-off-by: Cleber Rosa <crosa@redhat.com>