setup: Add default install_requires
Currently Avocado is unusable after pip installation as it requires
additional python libraries. Let's define the basic set in setup.py.

I did not included all dependencies, only the "default" set omitting
"fabric" which requires gcc to be installed and some "avocado.utils" and
restclient dependencies, which are not that frequently used.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>