runner: Try to interrupt test gently first
When the test does not produces the early_status in 60s or when the test
pushes the final status and not exits we SIGKILL the test, leaving us
with no trace on where it might hang. Let's use SIGTERM first and wait
for 1s to see whether the process won't be able to give us traceback,
otherwise nuke it with SIGKILL as before.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>