loader: keep the class discovery order
Currently we keep the order of discovery for the methods inside the
classes, but we don't do the same for the classes inside the module. The
consequence is that the order of the classes is given by the python
dictionary default sort, based on the hash of the dictionary items.

This patch makes the loader to use an `OrderedDict` during the discovery
process, making the order of the classes to be kept as they are
discovered.

Signed-off-by: Amador Pahim <apahim@redhat.com>