loader: recursive test discovery
This patch adds to the loader the ability to recursively discover tests
when the docstring `:avocado: recursive` is specified in the test class.

This recursive discovery will walk all the way up in the recursion
chain, adding the parents test methods to the child class information.

Reference: https://trello.com/c/GtFlhcOj
Signed-off-by: Amador Pahim <apahim@redhat.com>