loader: recursive test discovery
This patch adds to the loader the ability to recursively discover tests
when the docstring `:avocado: recursive` is specified in the test class.

This recursive discovery will walk all the way to the base class, bug an
`:avocado: disable` docstring in one class of the chain will interrupt
the recursion.

Reference: https://trello.com/c/GtFlhcOj
Signed-off-by: Amador Pahim <apahim@redhat.com>