jobdata: Store Varianter as data-only
Currently we just pickle the varianter object and hope we'll be able to
load it with other Avocado version. It usually works, but is not really
scalable. This patch adds a way to serialize Varianter into a list of
dicts and serializes it using json, turning the unsafe objects to
strings. Then it adds a way to load Varianter status using this
serialized state.

This procedure is lossy as it always assumes we are in non-debug mode
and it replaces the variants with safe representations. As a benefit one
does not need the original plugins to be able to reuse/replay the
variants as Avocado loads the representation using fake classes.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>