Docs: only build API docs for optional plugins that have been setup
There's a problem with assuming that all existing plugins will
be setup at API docs creation time, initially on RPM builds but
could also happen on systems that lack dependencies for some of
the plugins.

Let's only build docs for those that have been properly setup
and are available to be imported.

Signed-off-by: Cleber Rosa <crosa@redhat.com>