varianter: Add `to_str` feature to FakeVariantDispatcher
The `to_str` method is used to list/visualize the variants. As the
structure is not kept, let's just list all variants along with their
params.

This is a simplified version of the `mux` plugin's `to_str` and it
ignores `summary` and `variants` as it's not really useful nowadays, but
we can improve that in the future, if necessary.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>