jobdata: Allow loading jobs before TreeNode used TreeEnvironment
Before Avocado-49.0 TreeNode used to use 2 dicts to store values and their
origins, but the latest TreeNodes are using TreeEnvironemnt, which stores
everything related to the environment. This hack goes through plugins and
if it detects the old TreeNode it replaces them all with the
TreeEnvironments, making it possible to be used with the latest
AvocadoParams.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>