etc: Correct plugin names in avocado.conf
Currently there are two types of plugin names, one is the fully
qualified plugin name, which is the namespace used by stevedore and the
other is the module name, which is the module which handles the code.
It's probably too late to adjust the behavior, but let's at least
provide correct examples to simplify the usage.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>