Run result event plugins from the test runner process
One thing that has slipped up to this day, is that some code intended
to be run from the test runner process is actually running on the test
process.

A better solution to the fake_early_status given here is to change
the interface of `start_test`, so that only a subset of information
(available before the test actually returns its early state) is
available.  Yet another alternative, is to only call `start_test`
after the test process has been started and has returned its
*real* early status.  Since the current timeout is huge (60s)
I think that *could* result in bad user experience.

Signed-off-by: Cleber Rosa <crosa@redhat.com>