HTML: add a link to the report in the same location other results live
All other result plugins write (by default) to the job results dir.
The HTML plugin was different because it used to require a number of
files.  Now, it's a single file, and thus, could/should live in the
same place as other files.

To ease the transition of the location, let's add a symlink, so that
users can start to rely on the new proposed location.

Signed-off-by: Cleber Rosa <crosa@redhat.com>