Broken plugin notification: fix behavior with additional config files
When trying out Avocado installations, I wanted to silence the plugin
errors that I knew would fail to load.  I chose to do that by running
a command such as:

 $ avocado --config=silent_plugins.conf run passtest.py

But it did not behave as expected.  The reason is that plugins are
loaded first, and then the parsing of command line (with additional
parsing of config files) is done.  Let's parse the command line
earlier then.

Signed-off-by: Cleber Rosa <crosa@redhat.com>