loader: Get full-mapping after discovery
Currently only unbounded mapping is possible as on failure we want to
list all the available tests by given loaders without initializing them
first. This patch allows second-stage mapping which is a bounded method
and is called after `LoaderProxy.discover`, allowing the test loaders to
update the mappings according to the test discovery.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>