optional_plugins: Add yaml_testsuite_loader
Let's add yaml_testsuite loader which allows parsing YAML file and
producing testsuite out of it.

At this point due to Avocado limitation it only supports
FileLoader-based test discovery, but it should be possible with greater
changes to extend the scope of this loader for just about any Avocado
loader.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>