test: Remove unnecessary LazyProperty from Test class
Most of the LazyProperty variables in Test are actually always read on
"get_state" and other occasions so it does not make sense to set it
Lazily.

Also protect the remaining lazily initialized property from overriding
by using @property instead of our custom LazyProperty, which is not
really a property.

This commit requires some adjustments to nasty example tests which used
to override self.srcdir, which is not allowed (and restricted since this
commit).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>