runner: fix SIGINT behaviour
According to GNU documentation, "...typing C-c sends the signal SIGINT
to terminate all the processes in the foreground process group". But
when the SIGINT in directly sent to the Avocado main process, the test
process receives nothing, making the main process to wait indefinitely
for the test process to finish.

This patch makes the runner to send a SIGINT to the test process when
the main process receives a SIGINT, replicating the behaviour we have
on C-c.

Reference: https://trello.com/c/5RPZPH4F
Signed-off-by: Amador Pahim <apahim@redhat.com>