runner: Add "[std*] " prefix to test's sys.stdout|stderr in test log
Writing to sys.stdout|stderr from inside test should be treated as any
other output in test and should be prefixed by [std* ] to emphasize it's
not comming directly from logging, but from direct output. This prefix
should  and is only be prefixed in job.log and not in stdout|stderr
files.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>