REMOVEME: Redirect 1 and 2 to logging
This is very broken commit, which demonstrates why is it not OK to do
such thing. Basically right now Avocado coworks on producing the output
in a way that:

main-avocado:
 * UI
 * job.log (regarding job messages)
 * some framework messages (debug)

test-process:
 * test output to debug.log
 * test output to job.log (complements the main-avocado's job)
 * test output to stdout (when enabled by --show test)
 * redirects sys.stdout to `avocado.test.stdout`
 * redirects sys.stdout to `avocado.test` with prefix [stdout]
 * ditto sys.stderr

Now this is not a clean solution, but so far it worked well. With this
commit the main-avocado stays the same, but test-process has additional
tasks:

 * dup(1) and dup(2) to reserve original stdout/err with a different
   fileno, because standard logging requires raw output to it
 * replace StreamHandlers in logging to use the dupped 1/2s (raw output)
 * create pipes for plain stdout/err and replace 1/2 with them
 * create threads that periodically read the 1/2 pipes and forward the
   output to the stdout/err redirect (the avocado.test and
   avocado.test.stdout with prefix)

This implementation is hackish, for some reason it kept hanging on
closed pipe (probably I forgot to close them somewhere) so I used the
list, just ignore that part and only see the workflows which are:

os.write(1, "foo")
 1. 1 is dupped to stdout_in pipe
 2. stdout_out pipe is read
 3. output is forwarded to output.LoggingFile
 4. LoggingFile pushes the message to avocado.test with "[stdout] "
    prefix
 5. avocado.test writes the message to dupped 1 fd, therefor the record
    is displayed along with log level ... (if avocado.test is enabled)
 6. avocado.test writes the message to job.log
 7. avocado.test writes the message to debug.log
 8. LoggingFile pushes the message to avocado.test.stdout
 9. avocado.test.stdout writes the message to stdout (no "[stdout] "
    prefix this time)

One simplification can be done when we start initializing the test
output dir from runner. Then we'd have everything available from runner
and we'd be able to "simply":

 * dup(1/2) to get raw_1 and raw_2
 * replace StreamHandlers with to use the raw_1/2 filenos
 * replace 1/2 with pipe_ins
 * start threads to forward pipe_outs to output.LoggingFile

As the 1/2 changes the `sys.stdout` and `sys.stderr` should
(theoretically point to the pipes and get the prefix automatically.

Anyway I believe this handling is becoming cumbersome and IMO not worth
the effort.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>