Fix signal handlers
The SIGINT/SIGTERM handlers are currently broken. When we execute a
SIMPLE TEST, which ignores SIGINT/SIGTERM, the test process will not
finish on SIGINT/SIGTERM, becoming orphan.

---

Reproducer::

    $ cat /tmp/test.py
    #!/usr/bin/env python
    import signal
    import time

    if __name__ == "__main__":
        signal.signal(signal.SIGINT, signal.SIG_IGN)
        signal.signal(signal.SIGTERM, signal.SIG_IGN)
        signal.signal(signal.SIGQUIT, signal.SIG_IGN)
        while True:
            time.sleep(0.1)

    $ avocado run /tmp/test.py

Then play with `Ctrl+C` (once and twice), with `kill -2 <pid>` and
`kill <pid>`. Check if Avocado hangs or whether any process is left
behind.

---

Also, the only test we have for those cases is supposed to catch the
issue with SIGINT (test_interrupt.InterruptTest.test_badly_behaved), but
it's not really working.

This patch fixes the SIGINT handler in the avocado.process.SubProcess,
making it to just forward the SIGINT to the subprocess, calling the
default_int_handler() only if the subprocess finishes with the SIGINT.

Also, this patch creates a new SIGTERM handler in the
avocado.process.SubProcess, which tries to finish the subprocess sending
the same SIGTERM to it. If that's not enough, as a last resort, the
SIGTERM handler kills the subprocess.

With those handlers in place, the runner can send a SIGTERM to the test
process instead of a SIGKILL when user sends SIGINT twice to the main
process (via kill or via Ctrl+C).

Also, a new SIGTERM handler was introduced in the runner to send to the
test process the same SIGTERM.

Tests were rewritten to make sure they are testing the cases and the
Ctrl+C test was unified with the SIGINT test, since they test pretty
much the same. Also, a new test for the SIGTERM case was created.

Signed-off-by: Amador Pahim <apahim@redhat.com>

Signed-off-by: Amador Pahim <apahim@redhat.com>