minimal porting of avocado.utils.process to Python 3
Add two more properties stdout_bytes and stderr_bytes to CmdResult;
the decoded stdout and stderr are computed lazily.

For now, the default encoding is 'utf-8', though a better choice is
probably a combination of PYTHONENCODING, PYTHONIOENCODING and
`sys.getdefaultencoding()` (Lukaš suggested changing 'ascii' to
'utf-8' too).  The encoding and error handler can be customized,
though.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>