YAML Loader: move it into its own plugin and package
The newly introduced YAML Loader plugin is currently hosted on the
same plugin (directory and module) as the varianter one.  While they
share a lot of code, they have different purposes.

Let's move the YAML loader into its own plugin, so that its
installation is also optional and the separation is clearer.

Signed-off-by: Cleber Rosa <crosa@redhat.com>