Fix Travis check
With the EOL of Ubuntu 12.04, Travis CI is now running our tests using
Ubuntu 14.04. With this change, our check started to fail.

Even though I'm not sure about the reasons, I could make it work by
using a newer version of Pillow and also installing libvirt-python from
pip, since it's not working in the virtualenv if installed from apt.

Signed-off-by: Amador Pahim <apahim@redhat.com>