avocado.Test: use new data dirs for output check record feature
This changes the output check record functionality, from using the
file-level directory only containing std{out,err}.expected files, to
one that respects file-level, test-level and variant-level
directories.

When checking if the output matches the expected one, it just uses
the new public get_data() method.  When recording output data, it
either uses the variant-level when a variant is being used, or the
test-level directory when a variant is not being used.

A tool that could be added on top of this change of behavior is one
that consolidates recorded output in higher-levels.  Suppose that a
test run with a thousand variants generate the same output.  It's a
waste to have one thousand files (not counting the directories) when a
single file at the file-level data directory would be enough.  But,
this can be implemented later, I suppose.

Signed-off-by: Cleber Rosa <crosa@redhat.com>