avocado/core/test.py: rename TestName to TestID
The current TestName class records more than a test name (according
to the Avocado specification).  It records, and by default, presents
the Test ID (__str__() and __repr__()).

The Test Name, again according to the Avocado specifications, is
contained within a Test ID, so it's not necessary to have two
different classes IMO.

Finally, the str_filesystem() method was adapted to be coherent with
the fact that it returns a file system representation of the Test ID,
not the Test Name.  It also became a property, to match the same use
of str_uid and str_variant.

Signed-off-by: Cleber Rosa <crosa@redhat.com>