avocado/core/test.py: properly document the datadir property
There's a clear difference about of what a test file, and a test
itself, is in Avocado.  The datadir attribute though, has not been
created with such a difference in mind.

The best solution would be to renamed it to something like
"file_datadir", but since it's test a public (and stable) API, it's
not going to be changed anytime soon.

So, for now, let's clarify the meaning of the datadir property.

Signed-off-by: Cleber Rosa <crosa@redhat.com>